charts,manifests: Update the manual olm metering manifests.
This updates our manual olm manifests for deploying Metering. The CatalogSourceConfig has been a depreciated object for a while now, so that's been replaced with using a ConfigMap-based CatalogSource implementation. Note: there's currently no ConfigMap that gets templated, but there is a ./hack/create-upgrade-configmap.sh script that achieves the same functionality that can be renamed/generalized in down the line.