test: Stop returning an error when running the gather testing artifacts script.
Occasionally, we get an error in the e2e suite that we're reading from a file that's already been closed. It appears there might be a race between reading from the stdout pipe and the cmd.Run() method. This would stop appending any read buffer errors to the overall function's error array and in most cases, capturing these errors provide very little value.