pkg/deploy: Add the empty openshift.io/node-selector annotation to the metering-operator namespace.
In the case where a cluster administrator has configured the `defaultNodeSelector` field for the cluster's `Scheduler` object, which is responsible for determining the placement of Pods in a cluster, there may be conflicts when a user also specifies a `nodeSelector` configuration for the metering-operator's operands in the `MeteringConfig` object. Instead, if we always set the `'openshift.io/node-selector'` annotation to be empty, meaning that we don't care where the Pods in this namespace get scheduled, the `defaultNodeSelector` won't be injected by the Scheduler and the user-provided `nodeSelector` configuration will work as intended.

Note: these changes are only geared towards manually deploying Metering manually (e.g. deploy everything ourselves) and deploying Metering through OLM manually. In order to install Metering through OperatorHub, you would still need to create/label/annotate the namespace yourself as a pre-requisite step to configuring Metering.