images: Disable the presto-prometheus by default for OCP installations.
We expose this functionality in 4.5, but there hasn't been much progress integrating this connector throughout the metering stack, and the current OSBS imagebuilder production binary is having a hard time building the Presto RHEL images that contain this connector. If we instead disable this functionality by default, then we won't attempt to load this connector on the metering side-of-things, which will unblock QE for the time being, until this new imagebuilder binary has reached production.