charts,images: Reconcile a ConfigMap that contains the service-ca token secret.
In 4.1, the /var/run/secrets/kubernetes.io/serviceaccount/service-ca.crt field was depreciated, and in 4.5 it was removed entirely. Now, we need to migrate to reconciling an empty ConfigMap with the service.beta.openshift.io/inject-cabundle=true annotation, which results in the service CA certificate bundle being automatically injected through one of the cluster's controllers. We can then mount that certificate into any of the Metering operand Pods that require access to that CA bundle.