charts,manifests: Add the 'name=metering-operator' label selector to expose metrics properly.
When the metering-ansible-operator is deployed, the operator-sdk attempts to create a ServiceMonitor to expose metrics that Prometheus can scrape from an endpoint that the Pod exposes. As a result of this, this is the Service that gets created:

```yaml
...
spec:
  clusterIP: 172.30.31.205
  ports:
  - name: http-metrics
    port: 8383
    protocol: TCP
    targetPort: 8383
  - name: cr-metrics
    port: 8686
    protocol: TCP
    targetPort: 8686
  selector:
    name: metering-operator
  sessionAffinity: None
  type: ClusterIP
```

The problem here, is that the service is targeting pods that match the `name: metering-operator` label and the metering-operator deployment creates pods with the `app: metering-operator` label.

This means we can simply add the `name: metering-operator` label to the deployment pod template, so any pods that get created can match that service label selector.

We also needed to expose the 8383 and 8686 container ports so Prometheus can properly scrap those metrics, such as `workqueue_work_duration_seconds_count`.