images: Add validation check that ensures the randomly generated cookie seed length matches 32 characters.
This duplicates the check in the reporting-operator authproxy secret helm charts, but it's much easier and quicker to validate it here before we start reconciling metering-related resources. This also cleans up some of the 'when' conditionals to ensure that we're not attempting to grab values from dictionaries that don't exist.