pkg/deploy: Ensure the MeteringConfig CRD exists and has a populated Status field.
We're currently running into a problem in the upgrade e2e suite where we ensure the MeteringConfig CRD exists before creating the MeteringConfig custom resource.
This has proven to be an inadequate check, as seen in the following error:

```bash
FATA[06-08-2020 11:24:41] Failed to deploy metering: failed to install metering through OLM: failed to create the MeteringConfig resource: the server could not find the requested resource (post meteringconfigs.metering.openshift.io)  error="failed to install metering through OLM: failed to create the MeteringConfig resource: the server could not find the requested resource (post meteringconfigs.metering.openshift.io)"
```

Despite verifying the MeteringConfig CRD exists, we can still see that the apiserver is failing to find the plural name version of that resource.
If we instead modified the logic to ensure that the .Status.AcceptedNames.Plural matches the 'meteringconfigs' string, we can safely assume that the MeteringConfig CRD exists and it's "ready".