images: Combine the hive metastore overrides dictionary with the meteringconfig_spec dictionary in defaults/main.yml.
We need to avoid using set_fact on the meteringconfig_spec file before we're finished processing any non-reconciliation tasks. If we added an additional task file that we include after the configure_hive_metastore.yml task file, override a variable in the meteringconfig_spec dictionary, we won't see that propogate to the rest of the role as we ran set_fact on this dictionary in a previous task file.