test/e2e: Update the e2e suite to read from YAML manifests instead of giant Go objects.
This would add a set of MeteringConfig custom resource YAML manifests in the test/e2e/testdata directory. We can decode those manifests into Go object which is what the rest of the e2e deployframework suite currently expects. This way, we reduce the size of the main_test.go significantly, and it's much easier to follow the configuration listed in the YAML manifest, as compared to a giant Go object.