[release-4.3] charts,manifests: Add support for upgrading between Metering CSV versions.
In the 4.5 release cycle, support was added to support upgrading between Metering CSV versions. We were asked to backport upgrade support to any release that made sense, which in this case would be 4.3/4.4 as 4.2 is marked as EOL once 4.5 GA is released.

The [`olm.skipRange`](https://docs.openshift.com/container-platform/4.4/operators/understanding_olm/olm-understanding-olm.html#olm-upgrades-replacing-multiple_olm-understanding-olm) CSV annotation adds the ability to upgrade between Metering CSV versions in a single `package.yaml`, which ART populates.

This file governs the substitution rules that ART is configured to read from. In this case, we're substituting the second x.y.z version from <4.3.0 reference, with the latest 4.4.z z-stream release.