charts,manifests: Bump the minimum kubernetes version in the Metering CSV manifests to 1.18.
For some reason, we've neglected bumping this CSV field for a couple of releases now, so it's lagging a bit behind. In any case, we still need to bump to at least 1.16 as that's when CRDs went GA. We'll most likely want to bump to 1.18 as we use vendor those clientsets/apis throughout the project.