images: Fix conditional when checking if the cluster-wide proxy has been configured.
The first conditional check was configured to check if the `cluster_proxy_config` variable was defined. This is problematic check as that variable is the registered output of querying for the OCP cluster-wide proxy object, and in the case of Ansible, a registered variable will always be defined even if a task has been skipped. If we, instead, check if the `.resources` dictionary key is defined instead of the top-level object, that's sufficient as a baseline check.