pkg/deploy: Update the label selector used to delete the metering clusterroles/clusterrolebindings.
When we're reconciling the clusterrole and clusterrolebinding resources for the reporting-operator, we add a pruning label for those resources:
```
apiVersion: rbac.authorization.k8s.io/v1
kind: ClusterRole
metadata:
  annotations:
    operator-sdk/primary-resource: tflannag/operator-metering
    operator-sdk/primary-resource-type: MeteringConfig.metering.openshift.io
  creationTimestamp: "2020-06-23T16:29:24Z"
  labels:
    app: reporting-operator
    metering.openshift.io/ns-prune: tflannag
    metering.openshift.io/prune: cluster-monitoring-view-rbac
```

These are namely hacks to ensure we cleanup all resources properly, but you can see that we tag the watched namespace and the resource with a `metering.openshift.io/*prune*` label to track these resources.

In the past implementation, we only searched for clusterroles and clusterrolebindings that met the `app=reporting-operator` label selector, but we should also be searching for that namespace pruning label to avoid deleting other tenant's resources.