olm_deploy/: Add an additional set of manifests used to build a local registry.
In order to deploy the most up-to-date operator bundle, we need to be able to create a CatalogSource custom resource that references an existing image.

We have a couple of different options for doing this:
- Create ConfigMap containing the package.yaml, CRDs, and CSV manifests.
- Build and push an image containing the manifest bundle and an initialized registry server to an image registry.
- Build and push an image containing the manifest bundle and serve the registry server locally.

In the case of CI, we need to take the third approach as they expose a series of environment variables pointing to the newly built metering/reporting operator images.

In the Dockerfile.registry, we copy over the current openshift manifest bundle to a known destination path in the container, and copy over a script responsible for
manipulating the CSV manifest with the images that get built by CI or provided by devs through the METERING_OPERATOR_IMAGE_*/REPORTING_OPERATOR_IMAGE_* variables.

In the manifest directory, we have a registry deployment, which is responsible for calling that CSV manipulation script we copied over in the Dockerfile.registry
in an initContainer, and serving that updated manifest bundle in a local registry server. We also have a registry service, which is responsible for exposing this
local registry and can be consumed by a CatalogSource custom resource using the ClusterIP of that service and the grpc port we defined.

This mainly mirrors the setup that cluster-logging and the cluster-resource-override-admission operators have already configured.
The only main differences with those implementation is that we instantiate and configure the CatalogSource and registry manifests in Go, instead of manipulating
and applying those manifests using a client.