test/e2e: Add a wrapper function around the top-level TestMain function.
This is needed in the case where you want to defer any test cleanup as os.Exit doesn't respect that call in TestMain.
If we instead create a wrapper function around TestMain and call that wrapper function in TestMain, then we can start
cleaning up any testing resources in the e2e suite, instead of doing all of that work as part of the `trap cleanup EXIT` call in hack/e2e.sh.

Link to article that explains that behavior:
- http://blog.englund.nu/golang,/testing/2017/03/12/using-defer-in-testmain.html