pkg/deploy: Stop erroring when a Metering namespace does not contain a CSV object.
This shouldn't be an error message for many of the situations this utility would get run, especially in the case of the e2e suite where we expect some installations will fail early and therefore a CSV would be generated. In the case of the CLI, a user may re-run the uninstall sub-command to clean-up non-default resources, like the ClusterRoles/ClusterRoleBinding resources, or the namespace that was provided.