test/reportingframework: Use context.Background for any of the report-related test workflows.
Ever since we bumped our k8s dependencies to 1.18, which now requires the explicit usage of the context library, I've seen the number of throlling errors increase quite a bit. When I was playing around switching the instances of context.Todo to context.Background, I saw a small decrease in these occurrences, so we should continue to monitor that and whether follow-up changes are needed.