test: Update the e2e suite to deploy Metering through OLM.
In order to deploy the most up-to-date operator bundle, we need to be able to create a CatalogSource custom resource that references an existing image.

We have a couple of different options for doing this:
- Create ConfigMap containing the `package.yaml`, CRDs and CSV manifests.
- Build and push an image containing the manifest bundle and an initialized registry server to an image registry.
- Build and push an image containing the manifest bundle and serve the registry server locally.

In the case of CI, we need to take the third approach as they expose a series of environment variables pointing to the newly built metering/reporting operator images.

In the `Dockerfile.registry`, we copy over the current openshift manifest bundle to a known destination path in the container and copy over a script responsible for manipulating the CSV manifest with the images that get built by CI or provided by devs through the `$METERING_OPERATOR_IMAGE_*`/`$REPORTING_OPERATOR_IMAGE_*` variables.

In the manifest directory, we have a registry deployment, which is responsible for calling that CSV manipulation script we copied over in the Dockerfile.registry in an `initContainer`, and serving that updated manifest bundle in a local registry server. We also have a registry service, which is responsible for exposing this local registry and can be consumed by a CatalogSource custom resource using the ClusterIP of that service and the grpc port we defined.

This mainly mirrors the setup that cluster-logging and the cluster-resource-override-admission operators have already configured.

The only main differences with those implementations are that we instantiate and configure the CatalogSource and registry manifests in Go, instead of manipulating and applying those manifests using a client.

This updates the `deployframework` object, creating another method for instantiating a local registry and a `CatalogSource` custom resource that points to that local registry. From there, we can update the installation closure used to create and manage an installation of metering, from the manual installation method to the existing OLM installation method.

In order to create this ` CatalogSource`, we need to first decode the YAML registry manifests into Go objects, which we can then further manipulate and manage ourselves before creating those resources using the core clientsets. For the registry deployment, there are a couple of things we need to update: the default registry image, which we currently don't have a default (e.g. quay.io/openshift/metering-registry) to point to the `--registry-image` Go test flag that was provided (which is configured to read from the `$IMAGE_METERING_ANSIBLE_OPERATOR_REGISTRY` is present). In the case where there's a non-empty `$METERING_OPERATOR_IMAGE_*` or `$REPORTING_OPERATOR_IMAGE_*` environment variable(s), we also update the values in the deployment to point to those overridden images.

A quick note on the `$METERING_OPERATOR_IMAGE_*` and `$REPORTING_OPERATOR_IMAGE_*` environment variables:
- CI exposes a `$IMAGE_FORMAT` environment variable, which we can further process by doing some substitution. We don't need to do this manipulation as we already have a hack/lib/test.sh script (which is sourced in the hack/e2e.sh script) that pulls out the repository and tag for the metering-ansible-operator and metering-reporting-operator images.