charts,manifests: Update the Report CRD to add a schema for 'spec.overwriteExistingData'.
It looks like this is a valid API field for the Report CRD, but wasn't added when we were initially adding the schema validation for this object. Before, when we were using apiextensions.k8s.io/v1beta1 CRDs, this was fine as pruning of unknown fields defaulting to false, but now that we transistioned to v1 CRDs, we need to add this field as a valid schema property.