test/e2e: Add support for running any pre-installation tests.
In order to add more flexibility to the e2e suite, we need to support creating any pre-installation resources before punting the current test context to the deploy-metering Go package. In order to do this we can create a function type that takes a deployframework.DeployerCtx object, with the assumption that when we run the preInstallFunc, that state has been initialized. This is useful in the case where we want to create things like an NFS provisioner, or create a secret containing the AWS credentials that will be used throughout the testing suite.