test: Update how upgrade testing is preformed to use the CatalogSource created in the beginning of the e2e run.
Before the recent changes made to the e2e suite that enables all Metering tests to be deployed through OLM, instead of the manual workflow, we needed to use a CatalogSource containing all of the operator's metadata to do upgrade testing.

Now, before runnign any top-level tests, we build up an image registry containing the most up-to-date (in the context of CI) metering-ansible-operator manifest bundle, and creating all the requisite resources to expose that bundle as a custom CatalogSource. If upgrade testing is enabled, we can expect that when we base installation tests have finished, this CatalogSource containing a newer payload will exist and we can edit the existing metering subscription to point to this CatalogSource.