test/e2e: Add an e2e test for configuring a MySQL instance for Hive metastore.
We recommend overriding the default, local embedded Derby database that Hive metastore is configured to use, with MySQL or PostgreSQL databases.

This would add a fairly straight-foward MeteringConfig YAML manifest that is configured to reference to a MySQL database instance in the `mysql` namespace. We create that namespace and the database instance using the `oc new-app ...` mysql 5.7 imagestream before firing off the metering installation responsible for using mysql as an underlying hive metastore database.