charts,manifests: Lower the minKubeVersion to 1.17.0 to unblock CI.
Right now, every PR is failing e2e as the minKubeVersion specified in our CSV resource is 1.18.3, which is greater than the server version (1.17.0) that the apiserver is advertising. In order to unblock CI, we can lower this value to 1.17.0 with the intention of reverting it once this issue is resolved.