charts,manifests: Remove the Ansible sidecar container.
In the 0.18.x operator-sdk release, the inotify-tools dependency was removed from the Ansible base image, which the Ansible sidecar container relies on for viewing Ansible logs.

We can instead just remove this sidecar container entirely as the operator container now has support for both the Ansible and operator logs coalesced.