test/e2e: Disable the node selector e2e test by default.
This test is causing frequent e2e flakes as we're adding this custom testing label
to the available sets of nodes in the prior to firing off any metering installations.
The result is that all of the metering installations fight for their pods to be scheduled,
which is fine as the machineautoscaler will provision new machines. With this node selector
test, that's problematic as the new machines that the machineautoscaler provisions will not
have the custom testing node label, so pods will never schedule.