charts,manifests: Add the zap-level argument to the ansible-operator container.
This field controls the internal ansible-operator logging level. We can default this logging level to logging everything with the intention of editing this value on the deployment/CSV objects if you want to change the logging type.

Other available values for this argument are 'debug' and 'error' (or providing the verbosity of the logs using an integer):
- https://github.com/operator-framework/operator-sdk/blob/be718958884e04b4c4adfab5e0054524bba5226b/pkg/log/zap/flags.go#L52

In the case where you only want to stream the Ansible playbook logs, and not the internal ansible-operator logs you would edit the value of this argument to '--zap-level=error' in the metering-ansible-operator deployment and that would reduce the general verbosity of the container logs.