test/e2e: Add a post-install test that verifies that the metering-ansible-operator is exposing metrics properly.
This is a follow-up to the fixes made in https://github.com/kube-reporting/metering-operator/pull/1235 to make sure we don't regress in future releases.

At the moment, all this post-install test does is exec into the `prometheus-k8s` Pod in the openshift-monitoring namespace and ensures that the workqueue_work_duration_seconds_count metric returns a successful status. The ansible-operator, which will expose this metric, creates a ServiceMonitor for the metering-operator-metrics service name, so we should be able to query for the value of this metric in the test namespace that has deployed Metering.