pkg/operator: Update the scheme used for the event recorder to use the meteringv1 scheme.
In order to broadcast k8s events to help report errors for the
metering-related resources, we need to use the metering scheme that gets
generated from the pkg/apis/metering/v1/register.go implementation
instead of the default scheme the controller was previously using the
report the leader election event. Changing the event recorder to use the
metering scheme, in addition to the default scheme allows us to
broadcast events to different object's lifecycles.