test/e2e: Migrate to using the 10.2 MariaDB imagestream in tests.
It looks like the MySQL 5.7 imagestream was recently removed from 4.6 clusters, and we still have the open BZ for Metering cannot work with the most recently GA'd MySQL release, which is causing this e2e test to constantly fail.

Looking at the MariaDB vs. MySQL compatibility chart, it looks like 10.2/10.3/10.4 are all compatible with MySQL 5.7:
- https://mariadb.com/kb/en/mariadb-vs-mysql-compatibility/

These changes just update the imagestream we create in the e2e test to use the mariadb:10.2 version.