bundle: Add operator manifest bundle for the new format.
The decision was made to migrate all optional operators from the old app-registry bundle format, to this new index image format, during the 4.6 release cycle.

Currently, ART and CI are configured to package our operator using the old manifest style and this PR introduces another set of bundle manifests that are currently a no-op.