charts,manifests: Re-generate the upstream CSV manifest to use the upstream logo.
This process was very similar to what was documented in [#1319](https://github.com/kube-reporting/metering-operator/pull/1319). Basically, just base64 encrypted that svg file, checked if that was displaying properly for me locally, and then proceeded to update the charts to use that base64 output. From there, I push an operator manifest bundle containing the upstream manifests, created a CatalogSource custom resource that pointed to that image, and verified that the logo was correctly substituted in the openshift console.