test: Upgrade from metering version that contains the 4.5 many-to-many fix.
We're seeing consistent upgrade failures due to the 4.5 metering-ansible-operator version in the redhat-operators CatalogSource custom resource not containing the PVC-related many-to-many fix. As a result, upgrade tests will fail as the pre-upgrade tests are reporting errors due to metrics not being correctly important (namely due to the aforementioned many-to-many bug), but the actual upgrade and tests against that upgrade are succeeding.