test/deployframework: Populate the tenancy labels before creating the registry-related resources.
We're currently running into the problem where multiple metering e2e runs are conflicting with the registry-related resources we're spinning up for developers. In the case of the service object that gets created, we're only selecting on a single Pod label, and in the case where there's multiple registry deployments that all use the same label selector, we risk running e2e tests against a different version of metering than what we expected. This updates the static label selectors we use for the registry deployment and service, adding a `tenancy` label selector that we will populate at runtime after decoding the YAML manifest into a Go object.