charts,manifests: Add the secretName property to the MeteringConfig CRD schema validation.
The functionality for configuring the underlying Hive Metastore database to use a secret containing the username and password credentials for a mysql/postgresql instance was added early in the 4.6 release cycle so this isn't a regression from previous releases. When this feature was added, it looks like this field wasn't added as a property to the schema definition, which means that any user-provided MeteringConfig custom resource configuration that attempts to specify this secretName property will be rejected as it's considered an invalid schema definition (as the field is undefined) when using v1 CRDs.

These changes add the secretName property to the parent `db` object and runs `make metering-manifests` so the rendered YAML manifests are up-to-date.