hack,test: Add a custom go test flag that points to an index image
Update the e2e.sh wrapper and the entrypoint to the e2e suite to consume an additional go test flag.

Note: in the case we're running e2e tests in a CI environment, both of the bundle and index images are ephemeral and won't be promoted. This means we cannot set a reasonable default value for the index image environment variable/flagset.