test: Deploy metering using the index image packaging
Makes minor changes in the way we expose the metering-ocp package that test installations can consume. Before, we spun up a local registry containing the old packagemanifest versioning. Now, with the introduction of index images and the depreciation of the packagemanifest/app-registry format, we can simply create a CatalogSource custom resource that points to an existing index image. We'll most likely need to continue iterating on this approach, and determine whether it makes sense to provide an avenue to deploy metering using both approaches. For now, this only uses the index image format.