test/e2e: Reduce reporting-operator CPU requests and rename metering tests
With the migration to building the metering bundle and index images in CI, we needed to switch to a workflow that has smaller default AWS instance sizes than what we're use to running in CI, even when machine autoscaling has been enabled.

In order to schedule the reporting-operator Pods, we can limit the CPU requests which allows for the reporting-operator to fit inside of a recently scaled up machine.

This also renames some of the test/e2e/main_test.go test table names, which are used as a test namespace suffix as those names didn't make much sense in that context.