cmd/reporting-operator: Minor changes to the main package
Swap around the runReporting function parameters to have the context.Context argument come first.

Remove an unneeded function wrapper around adding cobra.Command(s) to the root command set.

Avoid hardcoding the path in the container to the namespace the reporting-operator Pod has been deployed in with a global variable.