hack,images/metering-ansible-operator: Migrate to using helm 3.x for templating resources
Update the deploy_resources.yaml task file, which is responsible for templating helm charts into YAML manifests so we can pass that to the `k8s` Ansible module.

Update the hack/create-metering-manifests.sh bash script, replacing the `-x` flag with `-s` when templating individual resources.

This migration from helm 2.x to helm 3.x is fairly trivial as we only utilized helm for it's templating and nothing deployment (i.e. tiller) facing.

Note on the changed flags:
- The global `--name` flag was removed.
- The `--execute` (`-x`) was renamed to `--show-only` (`-s`) and the functionality remains the same.

Quick note on the testing that was done locally:
- Spun up metering using a custom built helm image that contains the helm release-3.0 branch.
- Editted the local MeteringConfig custom resource, replacing the storage configuration and changed some of the reporting-operator custom parameters.