test/e2e: Increase the reporting-operator CPU requests due to 504 errors in API
Bump the reporting-operator CPU requests in the MeteringConfig YAML manifests we run consistently.

I was seeing more frequent 504's being returned when attempting to query for the report results in tests, and saw some improvements in fixing that behavior by bumping the resources on the proxy side-of-things.

Note: this resources were previously set to `cpu: 1` but the CI default machine set size was too low to schedule all the requisite Metering Pods, even with autoscaling enabled. Our CI configuration has been updated to always use m5.xlarge instance sizes, so we can safely revert that reduction in resources.