Dockerfile: Use the upstream ansible-operator base image
Update the dev Dockerfile.metering-ansible-operator to use the upstream operator-framework/ansible-operator base image instead of the downstream fork.

Before, you were unable to build this image locally due to the push to using RHEL8 images downstream, but this Dockerfile is used solely for local developmental purposes and isn't used in the CI/ART pipelines.

The operator-framework/ansible-operator Dockerfile uses UBI8 as their base image, so `pip3` usage is now required.