test/e2e: Add a post-install test that ensures the Metering leader election ConfigMap is present
Add a new post-install test that ensures that a ConfigMap is present that is used to determine the Metering Operator leader election.

Update the test/e2e/main_test.go e2e entrypoint function, adding this post-install function to the list of tests that will run against the static data input Metering installation.