images: Map the `meta` Ansible version to `ansible_operator_meta` in the role
Update the watches.yaml file, which contains the API definition for resources that the Metering Operator is responsible for watching and ensure that the `meta` Ansible variable maps to `ansible_operator_meta`.

Before, the `meta` variable was exposed through the ansible-operator and you would be able to pull information about the `MeteringConfig` custom resource that it's watching via `{{ meta.namespace }}` or `{{ meta.name }}`. Now, with the ansible-operator v1 version, that variable has been removed in favor of `ansible_operator_meta`. Explicitly mapping that variable to the `meta` Ansible variable ensures the same behavior we expect in previous ansible-operator versions.