olm_deploy: Remove the unused scripts directory
Remove the unused init-bundle-registry.sh bash script that is no longer needed (and the parent scripts directory).

Before, we needed to do some CSV manipulation at CI runtime which resulted in needed this script to substitute the images that were built by CI in favor of the default one listed in the metering CSV YAML manifest.

Now, the ci-operator has the ability to create bundle and index images for us which results in less workarounds needed.