hack,Makefile: Add support for mirroring changes to the metering bundle manifests
Update the Makefile to expose the current release of Metering. This so far is only used in one place, but introducing this variable can be propagated to the other various hack scripts used throughout metering.

Update the hack/generate-metering-manifests.sh bash script to mirror any changes made to the manifests/deploy/openshift/olm/bundle/<version> to the olm_deploy/bundle artifacts. The former is used downstream by ART to build the operator and the latter is used by CI to generate a bundle (and an index image containing that bundle image) for the operator that gets deployed at runtime.