test/e2e: Avoid using the fatal logrus level in the main_test.go
The logrus.Fatal logging level logs an entry at the fatal level and then calls os.Exit(1) after logging. This is problematic in the case of main_test.go where we expect that a custom defer function will be called if something problematic happens after creating the registry resources. If we instead log that error to the error log level and then return 1, we can continue getting the same behavior, but the defer function will still be ran.