hack: Gather all the PackageManifests in a test cluster
Update the hack/gather-registry-artifacts.sh bash script, and gather the overall list of PackageManifest custom resources that are stored in the openshift-marketplace namespace.

During an individual CI run, the e2e suite is responsible for creating a CatalogSource custom resource that references an index image that the ci-operator creates before a cluster is created. When the corresponding PackageManifest custom resource is not created from that CatalogSource, then querying for the "catalog-namespace=openshift-marketplace,catalog=metering-dev" label selector will return a list of zero indices so this would provide additional debugging measures to help determine if there's something wrong on the metering bundle side-of-things, or OLM side-of-things.