images: Avoid vendoring the crpyto.py Ansible module utilities
Remove the module_utils (and as a result, the module_utils.crypto.py library) directory in the meteringconfig Role.

Previously, this was necessary in order to have Metering function on FIPS-enabled clusters as the crypto.py module was being utilized in the openssl-* Ansible modules. This library listed the md5 algorithm as FIPS-compliant, despite md5 not being FIPS-compliant in many FIPS standards.

Now, this behavior has been patched upstream in Ansible (in versions greater than 2.9.6) so vendoring this library is no longer needed.