Dockerfile: Ensure that Ansible 2.9.6+ is installed in the container
Update the metering-ansible-operator Dockerfile(s), ensuring that the minimum Ansible version is 2.9.6.

The 2.9.6 version contains the patched crypto.py library that previously needed to be vendored in order to have Metering run on FIPS-enabled clusters.

Note on the Dockerfile.metering-ansible-operator: by default, we inherit Ansible from the base image and there's problems with upgrading from Ansible 2.9 to 2.10 when using pip. In order to bypass that limitation, we can specify 2.9.6 as the lower bound and 2.10 as the upper bound. It likely doesn't matter what Ansible version we get, but we don't utilize any 2.10 functionality or haven't tested that release as-of-yet.